% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{.empty_plot}
\alias{.empty_plot}
\title{Build an empty ggplot with a centered message}
\usage{
.empty_plot(message)
}
\arguments{
\item{message}{Character. Message to display.}
}
\value{
A ggplot object.
}
\description{
Build an empty ggplot with a centered message
}
\keyword{internal}
